+++ mobile/android/android-components/components/browser/engine-gecko-nightly/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngineSession.kt
@@ -78,6 +78,8 @@ class GeckoEngineSession(
 
     private var initialLoad = true
 
+    private var requestFromWebContent = false
+
     override val coroutineContext: CoroutineContext
         get() = context + job
 
@@ -89,6 +91,7 @@ class GeckoEngineSession(
      * See [EngineSession.loadUrl]
      */
     override fun loadUrl(url: String) {
+        requestFromWebContent = false
         geckoSession.loadUri(url)
     }
 
@@ -96,6 +99,7 @@ class GeckoEngineSession(
      * See [EngineSession.loadData]
      */
     override fun loadData(data: String, mimeType: String, encoding: String) {
+        requestFromWebContent = false
         when (encoding) {
             "base64" -> geckoSession.loadData(data.toByteArray(), mimeType)
             else -> geckoSession.loadString(data, mimeType)
@@ -316,9 +320,12 @@ class GeckoEngineSession(
             } else {
                 notifyObservers {
                     // Unlike the name LoadRequest.isRedirect may imply this flag is not about http redirects. The flag
-                    // is "true if and only if the request was triggered by user interaction."
+                    // is "True if and only if the request was triggered by an HTTP redirect."
                     // See: https://bugzilla.mozilla.org/show_bug.cgi?id=1545170
-                    onLoadRequest(triggeredByUserInteraction = request.isRedirect)
+                    onLoadRequest(
+                        triggeredByRedirect = request.isRedirect,
+                        triggeredByWebContent = requestFromWebContent
+                    )
                 }
 
                 GeckoResult.fromValue(AllowOrDeny.ALLOW)
@@ -386,6 +393,10 @@ class GeckoEngineSession(
         }
 
         override fun onPageStop(session: GeckoSession, success: Boolean) {
+            // by the time we reach here, any new request will come from web content.
+            // If it comes from the chrome, loadUrl(url) or loadData(string) will set it to
+            // false.
+            requestFromWebContent = true
             notifyObservers {
                 onProgress(PROGRESS_STOP)
                 onLoadingStateChange(false)