+++ mobile/android/android-components/components/browser/engine-system/src/main/java/mozilla/components/browser/engine/system/SystemEngineView.kt
@@ -226,7 +226,7 @@ class SystemEngineView @JvmOverloads constructor(
             }
 
             if (request.isForMainFrame) {
-                session?.let { it.notifyObservers { onLoadRequest(request.hasGesture()) } }
+                session?.let { it.notifyObservers { onLoadRequest(request.hasGesture(), true) } }
             }
 
             return super.shouldInterceptRequest(view, request)