+++ mobile/android/android-components/components/browser/session/src/main/java/mozilla/components/browser/session/engine/EngineObserver.kt
@@ -8,6 +8,8 @@ import android.graphics.Bitmap
 import android.os.Environment
 import mozilla.components.browser.session.Download
 import mozilla.components.browser.session.Session
+import mozilla.components.browser.session.engine.request.LoadRequestOption
+import mozilla.components.browser.session.engine.request.plus
 import mozilla.components.concept.engine.EngineSession
 import mozilla.components.concept.engine.HitResult
 import mozilla.components.concept.engine.manifest.WebAppManifest
@@ -41,10 +43,20 @@ internal class EngineObserver(
         }
     }
 
-    override fun onLoadRequest(triggeredByUserInteraction: Boolean) {
-        if (triggeredByUserInteraction) {
+    override fun onLoadRequest(triggeredByRedirect: Boolean, triggeredByWebContent: Boolean) {
+        if (triggeredByWebContent) {
             session.searchTerms = ""
         }
+        var triggers = LoadRequestOption.NONE.toMask()
+        if (triggeredByRedirect) {
+            triggers += LoadRequestOption.REDIRECT
+        }
+
+        if (triggeredByWebContent) {
+            triggers += LoadRequestOption.WEB_CONTENT
+        }
+
+        session.loadRequestTriggers = triggers
     }
 
     override fun onTitleChange(title: String) {