+++ mobile/android/android-components/components/browser/session/src/test/java/mozilla/components/browser/session/SessionTest.kt
@@ -10,6 +10,8 @@ import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.async
 import kotlinx.coroutines.runBlocking
 import mozilla.components.browser.session.Session.Source
+import mozilla.components.browser.session.engine.request.LoadRequestOption
+import mozilla.components.browser.session.engine.request.isSet
 import mozilla.components.browser.session.tab.CustomTabConfig
 import mozilla.components.concept.engine.HitResult
 import mozilla.components.concept.engine.manifest.WebAppManifest
@@ -159,6 +161,20 @@ class SessionTest {
 
         assertEquals("mozilla android", session.searchTerms)
         verify(observer, times(1)).onSearch(eq(session), eq("mozilla android"))
+        verifyNoMoreInteractions(observer)
+    }
+
+    @Test
+    fun `observer is notified when load request is triggered`() {
+        val observer = mock(Session.Observer::class.java)
+
+        val session = Session("https://www.mozilla.org")
+        session.register(observer)
+
+        session.loadRequestTriggers = LoadRequestOption.REDIRECT.toMask()
+
+        assertTrue(session.loadRequestTriggers.isSet(LoadRequestOption.REDIRECT))
+        verify(observer, times(1)).onLoadRequest(eq(session), eq(true), eq(false))
         verifyNoMoreInteractions(observer)
     }