+++ mobile/android/android-components/components/browser/session/src/test/java/mozilla/components/browser/session/engine/EngineObserverTest.kt
@@ -6,6 +6,8 @@ package mozilla.components.browser.session.engine
 
 import android.graphics.Bitmap
 import mozilla.components.browser.session.Session
+import mozilla.components.browser.session.engine.request.LoadRequestOption
+import mozilla.components.browser.session.engine.request.isSet
 import mozilla.components.concept.engine.EngineSession
 import mozilla.components.concept.engine.EngineSessionState
 import mozilla.components.concept.engine.HitResult
@@ -445,19 +447,30 @@ class EngineObserverTest {
         session.searchTerms = "Mozilla Foundation"
 
         val observer = EngineObserver(session)
-        observer.onLoadRequest(triggeredByUserInteraction = true)
+        observer.onLoadRequest(triggeredByRedirect = true, triggeredByWebContent = true)
 
         assertEquals("", session.searchTerms)
+        val triggeredByRedirect = session.loadRequestTriggers.isSet(LoadRequestOption.REDIRECT)
+        val triggeredByWebContent = session.loadRequestTriggers.isSet(LoadRequestOption.WEB_CONTENT)
+
+        assertTrue(triggeredByRedirect)
+        assertTrue(triggeredByWebContent)
     }
 
     @Test
-    fun `onLoadRequest does not clear search terms for requests not triggered by user interaction`() {
+    fun `onLoadRequest does not clear search terms for requests not triggered by user interacting with web content`() {
         val session = Session("https://www.mozilla.org")
         session.searchTerms = "Mozilla Foundation"
 
         val observer = EngineObserver(session)
-        observer.onLoadRequest(triggeredByUserInteraction = false)
+        observer.onLoadRequest(triggeredByRedirect = true, triggeredByWebContent = false)
 
         assertEquals("Mozilla Foundation", session.searchTerms)
+
+        val triggeredByRedirect = session.loadRequestTriggers.isSet(LoadRequestOption.REDIRECT)
+        val triggeredByWebContent = session.loadRequestTriggers.isSet(LoadRequestOption.WEB_CONTENT)
+
+        assertTrue(triggeredByRedirect)
+        assertFalse(triggeredByWebContent)
     }
 }